requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Date/Time
python-dateutil==2.8.2

# File Upload
aiofiles==23.2.1
Pillow==10.2.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Utilities
configparser==6.0.0
boto3==1.34.34
